:truck: Refactor anemoi datasets and lightning dataloaders into a common folder
