+++ ice_station_zebra/data/anemoi/preprocessors/icenet_sic.py
